Done programming for the evening.
Current code on https://lc-abq-biopark.shinyapps.io/lc-abq-biopark/
Not throwing any errors. 
Very basic and needs a few more hours of work. 

Making the text based visuals in the UI file. The only thing the server.r does is upload the csv file. not much being tracked but I will as soon as I add the basic UI requirements on the visual page.